AFI copy/requirements.txt
fastapi==0.111.0
uvicorn==0.20.0
pandas==2.2.2
numpy==1.26.4
TA-Lib==0.4.28
ccxt==4.3.41
PyYAML==6.0.1
websockets==12.0
aiohttp==3.9.5
python-telegram-bot==21.2
pydantic==2.7.4
